refactor: tidied up sign in and sign out, still need to sort out limiting access when signed out
