+++ src/components/Authentication/firebase_hosting/SignOut.tsx
